requirements.txt
pytest
python-docx
black
python-dotenv
azure-functions
azure-identity
azure-storage-blob
